move normalization to a separate package
